refactor cache modules (#392)
