+++ neurons/validator/services/data_service.py
@@ -350,12 +350,12 @@ class DataService:
 
             try:
                 # Only count source types the uploader actually handles (miner verified + generated)
-                miner_img = self.content_manager.content_db.count_unuploaded_media(modality="image", source_type="miner")
-                gen_img = self.content_manager.content_db.count_unuploaded_media(modality="image", source_type="generated")
+                miner_img = self.content_manager.media.count_unuploaded_media(modality="image", source_type="miner")
+                gen_img = self.content_manager.media.count_unuploaded_media(modality="image", source_type="generated")
                 image_count = miner_img + gen_img
 
-                miner_vid = self.content_manager.content_db.count_unuploaded_media(modality="video", source_type="miner")
-                gen_vid = self.content_manager.content_db.count_unuploaded_media(modality="video", source_type="generated")
+                miner_vid = self.content_manager.media.count_unuploaded_media(modality="video", source_type="miner")
+                gen_vid = self.content_manager.media.count_unuploaded_media(modality="video", source_type="generated")
                 video_count = miner_vid + gen_vid
                 image_thresh = self.config.upload_image_threshold
                 video_thresh = self.config.upload_video_threshold
