+++ tests/test_challenge_outcomes.py
@@ -2,16 +2,25 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from gas.cache.content_db import ContentDB
+from gas.cache.db import ConnectionManager, PromptStore, MediaStore, ChallengeStore
+from gas.cache.db.connection import create_schema
 from gas.types import MediaType, Modality, SourceType
 
 
 class ChallengeOutcomeStatsTest(unittest.TestCase):
     def test_challenge_outcomes_count_pre_storage_failures(self):
         with tempfile.TemporaryDirectory() as tmpdir:
-            db = ContentDB(Path(tmpdir) / "prompts.db")
-            prompt_id = db.add_prompt_entry("a test prompt", modality="image")
-            media_id = db.add_media_entry(
+            db_path = Path(tmpdir) / "prompts.db"
+            conn = ConnectionManager(db_path)
+            with conn.connect() as c:
+                create_schema(c)
+
+            prompts = PromptStore(conn)
+            challenges = ChallengeStore(conn)
+            media = MediaStore(conn, prompts, challenges)
+
+            prompt_id = prompts.add_prompt_entry("a test prompt", modality="image")
+            media_id = media.add_media_entry(
                 prompt_id=prompt_id,
                 file_path=str(Path(tmpdir) / "media.png"),
                 modality=Modality.IMAGE,
@@ -23,7 +32,7 @@ class ChallengeOutcomeStatsTest(unittest.TestCase):
                 task_id="task-verified",
             )
 
-            db.record_challenge_outcome(
+            challenges.record_outcome(
                 task_id="task-verified",
                 uid=1,
                 hotkey="hotkey-1",
@@ -32,7 +41,7 @@ class ChallengeOutcomeStatsTest(unittest.TestCase):
                 status="verified",
                 media_id=media_id,
             )
-            db.record_challenge_outcome(
+            challenges.record_outcome(
                 task_id="task-rejected",
                 uid=1,
                 hotkey="hotkey-1",
@@ -42,7 +51,7 @@ class ChallengeOutcomeStatsTest(unittest.TestCase):
                 failure_reason="C2PA verification failed",
             )
 
-            stats = db.get_challenge_outcome_stats_last_n_hours(lookback_hours=1)
+            stats = challenges.get_outcome_stats_last_n_hours(lookback_hours=1)
 
             self.assertEqual(stats["hotkey-1"]["total_verified"], 1)
             self.assertEqual(stats["hotkey-1"]["total_failed"], 1)
